Repository refactoring
1) The structure of the repository has been changed. All python files has
been moved to the serversmanagemant python package.

2) All code from app/__init__.py file has been removed. Now all extensions
(such as SQLAlchemy, Alembic and others) are initialized in extensions.py
file. The app object is created using create_app function and  all
extensions initialize this object in create_app function.

3) create_db.py script was replaced with flask alembic and migrations.

4) Fixed code style according to PEP8. Added tox checks using flake8 and
hacking.

5) Changed the algorithm of choosing server to deploy environment. Now we
try to calc free RAM of the server using its total RAM (zabbix is used for
that), deployed environments on the server and the reserved by system RAM.
The least powerful server which have enough RAM and DISK is used.

6) Fixed some small problems such as deploying environment using not
jenkins user.

7) The direct import of views was replaced with blueprints usage.